move `virtualship init` logic to new initialise.py module
